requirements.txt
Jinja2==3.1.1
uvicorn==0.17.6
earthengine-api==0.1.306
fastapi==0.75.2
geopandas==0.11.0
python-decouple==3.6
PyYAML==6.0
pymongo==4.1.1
astropy==5.1
PyWavelets==1.3.0
matplotlib== 3.5.2
matplotlib-inline==0.1.3
scipy==1.8.1
pandas==1.4.2
numpy==1.22.3
pydantic~=1.9.0
orjson
git+https://github.com/andersonreisoares/stmetrics.git@phenology
